types and icons improvements
