+++ src/AwesomeApp.UI/example.env
@@ -9,5 +9,5 @@ AWESOME_API_KEY="TODO"
 NEXTAUTH_URL="https://localhost:3000"
 AWESOME_API_URL="https://localhost:7075"
 
-NEXT_PUBLIC_ENVIRONMENT="local"
+NEXT_PUBLIC_ENVIRONMENT="local | prod"
 NEXT_PUBLIC_API_URL="https://localhost:3000/api"
