+++ src/AwesomeApp.UI/src/backend/libs/ServerPropsHelpers.ts
@@ -3,15 +3,13 @@ import { getCsrfToken } from './CsrfToken'
 import { CsrfToken } from '@/common/types'
 import { HttpStatusCode } from 'axios'
 
-export const resultProps = <T>(props?: T, notFound?: boolean) => ({
-  notFound,
+type NotFound = { notFound: true }
+
+export const resultProps = <T>(props?: T) => ({
   props: props ?? { } as T
 })
 
-export const resultNotFound = <T>() => ({
-  notFound: true,
-  props: { } as T
-} as GetServerSidePropsResult<T>)
+export const resultNotFound = () => ({ notFound: true } as NotFound)
 
 export const resultRedirect = (destination: string) => ({
   redirect: {
@@ -31,7 +29,7 @@ export const resultStatusCode = <T>(context: GetServerSidePropsContext, statusCo
 
 export const resultForbidden = <T>(context: GetServerSidePropsContext) => resultStatusCode<T>(context, HttpStatusCode.Forbidden)
 
-export async function resultPropsWithCsrfToken<T>(context: GetServerSidePropsContext, props?: T, notFound?: boolean): Promise<{ props: T & CsrfToken | CsrfToken }> {
+export async function resultPropsWithCsrfToken<T>(context: GetServerSidePropsContext, props?: T): Promise<{ props: T & CsrfToken | CsrfToken }> {
   const authCookie = context.req.cookies['next-auth.session-token']
   
   if (!authCookie) {
@@ -39,8 +37,7 @@ export async function resultPropsWithCsrfToken<T>(context: GetServerSidePropsCon
       { 
         ...(props ?? {}), 
         csrfToken: undefined 
-      },
-      notFound
+      }
     )
   }
 
@@ -50,7 +47,6 @@ export async function resultPropsWithCsrfToken<T>(context: GetServerSidePropsCon
     {
       ...(props ?? {}),
       csrfToken
-    },
-    notFound
+    }
   )
 }
