+++ src/AwesomeApp.UI/src/frontend/views/app-nav-bar/styles.module.css
@@ -1,3 +1,7 @@
+.navBarButton {
+  text-transform: none;
+}
+
 .logoText {
   font-weight: 700;
 }
